requirements.txt
requests>=2.31.0
dnspython>=2.4.2
selenium>=4.15.2
beautifulsoup4>=4.12.2
webdriver-manager>=4.0.1
discord
selenium
pytz